コメント型とコンポーネントファイルの切り方を変更
